+++ amethyst_utils/src/removal.rs
@@ -1,10 +1,12 @@
 //! Provides utilities to remove large amounts of entities with a single command.
 
 use std::fmt::Debug;
+use std::ops::Deref;
 
 use amethyst_assets::PrefabData;
 use amethyst_core::specs::{
-    world::EntitiesRes, Component, DenseVecStorage, Entity, Join, ReadStorage, WriteStorage,
+    storage::MaskedStorage, world::EntitiesRes, Component, DenseVecStorage, Entity, Join, Storage,
+    WriteStorage,
 };
 use amethyst_derive::PrefabData;
 use amethyst_error::Error;
@@ -42,12 +44,13 @@ where
 }
 
 /// Removes all entities that have the `Removal<I>` component with the specified removal_id.
-pub fn exec_removal<I>(
+pub fn exec_removal<I, D>(
     entities: &EntitiesRes,
-    removal_storage: &ReadStorage<'_, Removal<I>>,
+    removal_storage: &Storage<'_, Removal<I>, D>,
     removal_id: I,
 ) where
     I: Debug + Clone + PartialEq + Send + Sync + 'static,
+    D: Deref<Target = MaskedStorage<Removal<I>>>,
 {
     for (e, _) in (&*entities, removal_storage)
         .join()